+++ firebase-dataconnect/CHANGELOG.md
@@ -10,6 +10,8 @@
   [#7833](https://github.com/firebase/firebase-android-sdk/pull/7833))
 - [changed] DataSource wired up and entity paths extracted from response.
   [#7814](https://github.com/firebase/firebase-android-sdk/pull/7814))
+- [changed] Offline caching now respects maxAge.
+  [#NNNN](https://github.com/firebase/firebase-android-sdk/pull/NNNN))
 
 # 17.1.3